+++ tests/test_geojson.py
@@ -1,5 +1,10 @@
+import json
+import decimal
+
 import pytest
+from numpy.ma.testutils import approx
 from shapely.geometry import shape
+from clldutils.jsonlib import load
 
 from cldfgeojson.geojson import *
 
@@ -17,3 +22,20 @@ from cldfgeojson.geojson import *
 )
 def test_pacific_centered(type_, coords, check):
     assert check(shape(pacific_centered(dict(type=type_, coordinates=coords))))
+
+
+def test_dumps(tmp_path):
+    with pytest.raises(TypeError):
+        dumps(decimal.Decimal('1'))
+
+    c = 1.23456789
+    assert '1.2345678' in json.dumps(c)
+    obj = json.loads(dumps({'properties': {'a': c}, 'coordinates': [[c], c]}))
+    assert approx(obj['properties']['a'], c)
+    assert approx(obj['coordinates'][0][0], 1.23456)
+
+    p = tmp_path / 'test.geojson'
+    dump({'properties': {'a': c}, 'coordinates': [[c], c]}, p)
+    obj = load(p)
+    assert approx(obj['properties']['a'], c)
+    assert approx(obj['coordinates'][0][0], 1.23456)